requirements.txt
python-telegram-bot==13.11
Pillow==9.0.1
qrcode===7.3.1
